Combine rewrite passes
