+++ src/targets/ref/target.cpp
@@ -4,8 +4,7 @@
 #include <migraphx/register_target.hpp>
 #include <migraphx/pass.hpp>
 #include <migraphx/auto_contiguous.hpp>
-#include <migraphx/rewrite_dequantizelinear.hpp>
-#include <migraphx/rewrite_quantizelinear.hpp>
+#include <migraphx/rewrite_quantization.hpp>
 #include <migraphx/rewrite_rnn.hpp>
 #include <migraphx/dead_code_elimination.hpp>
 #include <migraphx/generate.hpp>
@@ -20,9 +19,7 @@ std::string target::name() const { return "ref"; }
 std::vector<pass> target::get_passes(migraphx::context&, const compile_options&) const
 {
     return {normalize_ops{},
-            rewrite_dequantizelinear{},
-            dead_code_elimination{},
-            rewrite_quantizelinear{},
+            rewrite_quantization{},
             dead_code_elimination{},
             rewrite_rnn{},
             dead_code_elimination{},
